TRIVIAGEN-14: Moved points to viewmodel
