+++ app/src/main/java/com/triviagenai/triviagen/trivia/domain/model/TriviaQuestion.kt
@@ -5,8 +5,7 @@ import com.triviagenai.triviagen.trivia.data.model.Round
 data class TriviaQuestion(
     val question: String,
     val options: List<String>,
-    val answer: String,
-    val points : Int = 5,
+    val answer: String
 )
 
 fun Round.TriviaRound.mapToTriviaQuestions() = questions.map { question ->
