+++ app/src/main/java/com/triviagenai/triviagen/trivia/domain/repository/TriviaRepository.kt
@@ -1,6 +1,6 @@
 package com.triviagenai.triviagen.trivia.domain.repository
 
-import com.triviagenai.triviagen.trivia.data.model.Resource
+import com.triviagenai.triviagen.core.data.model.Resource
 import com.triviagenai.triviagen.trivia.data.model.Round
 import kotlinx.coroutines.flow.Flow
